Convert spaces to tabs
The inform compiler seems to let us get away with spaces for
indentation, but only sometimes, in some cases.  For nested
conditionals, it seems to rather require tabs.  So this commit
converts all _meaningful_ indentation to tabs for consistency.  In
some places, text strings are continued onto the next line using
spaces to indent.