Fix asking about things in other rooms
I finally discovered how to "allow access" for referring to distant
objects/people/etc.  This required some refactoring of "name badge"
afterward, since "name" is part of that and the parser found it
ambiguous when you could be referring to a general "name" concept or a
distant "name badge" object.