Refactor intro and day 1/2 scenes
If we want to set things up when day 1 / day 2 scenes begin, and treat
them as proper scenes, then ideally the intro text should happen
before any scene.  But attaching it to "when play begins" caused
ordering issues with the first scene, which also starts "when play
begins".  So now the intro text is not in any scene, but in a special
rule inserted early in the startup rulebooks.