Refactor game endings
Now we can say something like "show ending 1" and get a consistent
printout showing which ending the player found, how many endings there
are, and how many turns they took to get there.

Factoring the printout into an activity that takes a variable ending
number caused a weird bug in the HTML output that I assume must be a
Vorple bug or a bug in the Vorple Screen Effects extension.  To work
around it, we don't use "center" in the source for this, and instead
center the output in CSS.  The terminal version will not be centered,
which is fine.