Fix scientist names
Dr. von Braun was being referred to as something random in the room
description (such as "north" or "nothing") because I used "[the noun]"
in context of a room description paragraph instead of the doctor
himself.  This is now fixed.

I also attempted to fix some disambiguation issues between the doctor
and the other scientists in the room, but this is still a work in
progress.  It's not quite right yet.  (See "test others" for the test
script which isn't giving me the right responses yet.)