Finally fix the phrase "other scientists"
I can't, for the life of me, understand why this was an issue.  If
"other scientists" is the exact name of the character, why wouldn't
that match that character instead of a similarly named character, "the
head scientist".

The hacky workaround I found is to edit the player's commands and
replace "other scientists" with "the others", which is an alias that
maps to the "other scientists" character.  It's so ridiculous that we
would need to do something like this.

Graham Nelson, if you can hear this, please save us from this madness!