Switch docker images for Inform
Most of my issues with GitHub Actions were really issues with the
Docker image I had chosen.  This one is more similar to my home setup,
and it seems to work.

Here's hoping I don't have to push another update to GitHub Actions to
get the automatic publication to gh-pages to work.