Tweak propulsion lab
The new material for the lab conflicts with some of the old material,
so this adds a TODO comment to resolve how the old puzzle fits with
this new characterization of the lab.

Dr. von Braun's content has been grouped together, moving the
chalkboards up.

The cage and tapir are now things in the room, and are highlighted.
After the player gets told the tapir is an aardvark, the name changes,
using the rules for strangers.

A couple of compilation errors are fixed on the rules for asking about
the tapir.

Finally, some small changes are made to the severity of Dr. von
Braun's accent, for consistency.