Refactor events and timed hints
Now there is an event class that simplifies the logic of dealing with
turn counting and timed events and hints.  This improves the hints
given in the propulsion lab, such that they will only start several
turns after entering the room, rather than based on turn count mod 5.
This also adds a periodic hint for the blueprints puzzle that mirrors
the one given for the propulsion lab.